Separate Gateway and Node tests
